step9 : user를 관리하기 위한 UserList 클래스 생성
