--- src/main/java/src/User.java
@@ -1,22 +1,0 @@
-package src;
-
-public class User {
-    private Integer life;
-    private Integer gameNum;
-
-    public Integer getLife() {
-        return life;
-    }
-
-    public void setLife(Integer life) {
-        this.life = life;
-    }
-
-    public Integer getGameNum() {
-        return gameNum;
-    }
-
-    public void setGameNum(Integer gameNum) {
-        this.gameNum = gameNum;
-    }
-}
